llm/corrective_rag/service/requirements.txt
pypdf==4.1.0
langchain==0.1.13
langgraph==0.0.30
langchain_community==0.0.29
langchain-text-splitters==0.0.1
langchain-openai==0.1.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
langchainhub==0.1.15
tiktoken==0.6.0
importlib-metadata==7.1.0
flask==3.0.2
waitress==3.0.0